pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "carbon_tracker"
version = "0.1.0"
description = "Carbon Tracker Agent for Road Freight Logistics"
requires-python = ">=3.8"
dependencies = [
    "pandas",
    "numpy",
    "scikit-learn",
    "streamlit",
    "plotly",
    "openai",
    "google-generativeai",
]

[tool.setuptools.packages.find]
where = ["."]

